+++ CHANGELOG.md
@@ -31,6 +31,7 @@
 - The configuration menu no longer has a reload button. This is done automatically.
 
 ### Fixed
+- Fixed the brush UI not showing the translated brush shapes and types.
 - Fixed the Clean type doing the opposite of what it is supposed to do.
 - Fixed the Leafblower type not dropping items as expected, and fixed its complete failure when used in a size that
   would make it executed asynchronous.